Fixed crashes in SSAO and CharacterAnimation sample
A d3d12 update made visible some errors that did not cause crashes
before (at least on some NVIDIA hardware).  SSAO requires an SRV to the
application depth buffer.  However, the depth buffer resource format was
specified as DXGI_FORMAT_D24_UNORM_S8_UINT.  Therefore, it was an error
to create an SRV to this resource with format
DXGI_FORMAT_R24_UNORM_X8_TYPELESS.  The fix is to create the depth
buffer resource with the typless format DXGI_FORMAT_R24G8_TYPELESS so
that we can create a DSV with format DXGI_FORMAT_D24_UNORM_S8_UINT and
an SRV with format DXGI_FORMAT_R24_UNORM_X8_TYPELESS.

Also fixed a small error with the SSAO normal map resource not being in
the expected initial state.